+++ Chauffeur.TestingTools/MockTextWriter.fs
@@ -1,4 +1,4 @@
-﻿namespace Chauffeur.Tests.Integration
+﻿namespace Chauffeur.TestingTools
 open System.Diagnostics
 open System.IO
 open System.Threading.Tasks
@@ -19,19 +19,3 @@ type MockTextWriter() =
         Debug.WriteLine value
         messages <- value :: messages
         Task.FromResult value :> System.Threading.Tasks.Task
-
-type MockTextReader() =
-    inherit TextReader()
-
-    let mutable remainingCommands = []
-
-    member x.AddCommand command =
-        remainingCommands <- remainingCommands @ [command]
-
-    override x.ReadLineAsync() =
-        match remainingCommands with
-        | [] -> failwith "No commands"
-        | head::tail ->
-            remainingCommands <- tail
-            head
-        |> Task.FromResult