+++ Chauffeur.Tests.Integration/TestHelpers.fs
@@ -1,26 +1,5 @@
 ﻿module TestHelpers
 
-open System
-open System.IO
-open System.Reflection
-open Chauffeur.Tests.Integration
-open Chauffeur.Host
-
-let private cwd = FileInfo(Assembly.GetExecutingAssembly().Location).Directory.FullName
-let private dbFolder = "databases"
-
-let private setDataDirectory() =
-    let now = DateTimeOffset.Now
-    let ticks = now.Ticks.ToString()
-
-    let folderForRun = Path.Combine [|cwd; dbFolder; ticks|]
-
-    Directory.CreateDirectory folderForRun |> ignore
-
-    AppDomain.CurrentDomain.SetData("DataDirectory", folderForRun)
-
-    folderForRun
-
 let knownTables =
     ["cmsContent";
     "cmsContentType";
@@ -67,28 +46,10 @@ let knownTables =
     "umbracoUser2app";
     "umbracoUser2NodeNotify";
     "umbracoUser2NodePermission";
+    "umbracoUser2UserGroup";
+    "umbracoUserGroup";
+    "umbracoUserGroup2App";
+    "umbracoUserGroup2NodePermission";
+    "umbracoUserStartNode";
     "umbracoUserType";
     "umbracoLock"]
-
-[<AbstractClass>]
-type UmbracoHostTestBase() =
-    let dbFolder = setDataDirectory()
-
-    let writer = new MockTextWriter()
-    let reader = new MockTextReader()
-    let host = new UmbracoHost(reader, writer)
-
-    member x.DatabaseLocation = dbFolder
-    member x.Host = host
-    member x.TextReader = reader
-    member x.TextWriter = writer
-
-    interface IDisposable with
-        member x.Dispose() =
-            writer.Dispose()
-            reader.Dispose()
-            host.Dispose()
-
-let getChauffeurFolder databaseLocation =
-        let chauffeurFolder = Path.Combine [| databaseLocation; "Chauffeur" |]
-        Directory.CreateDirectory chauffeurFolder