+++ Chauffeur/TextWriterExtensions.cs
@@ -5,7 +5,7 @@ using System.Reflection;
 using System.Threading.Tasks;
 namespace Chauffeur
 {
-    static class TextWriterExtensions
+    public static class TextWriterExtensions
     {
         public static async Task WriteLineFormattedAsync(this TextWriter writer, string format, params object[] arguments)
         {